Gradient in linear form
This is what the color stops would be in linear form. Good luck
wrapping it around half a circle.

The way I did it in Illustrator was to divide the circle up into
1/10ths, and do a linear gradient from one inside vertex to the other
inside vertex for each arc section, which did a reasonable
approximation.

Sorry I did bad things to the repo. :(